requirements/app.txt
galaxy.plugin.api==0.30
python-dateutil==2.8.0
requests==2.21.0
psutil==5.6.1
PyYAML==5.1
